pyproject.toml
[project]
name = "VietnameseAlphabetPronunciation_DataCollection"
version = "0.1.0"
description = "VietnameseAlphabetPronunciation_DataCollection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dropbox>=12.0.2",
    "streamlit>=1.41.1",
    "streamlit-audiorecorder>=0.0.6",
]
